Make gained EXP a 32bit value
This was a lot more complicated than I thought.
Gamefreak loves two things.
1. copying variables byte-by-byte into random buffers and
2. copying code between files

This commit deals with both of those as apparently the entire XP system
exists twice for some reason, and XP values are passed through no fewer
than two buffers, one of which was very clearly not designed to hold
32bit values at all, ever.