Set game version when opening start menu
Previously, these used to be in the main game loop.
Moving them out of that didn’t actually work because they were set
before the savegame was even loaded.
In the start menu, they’re set whenever needed (because the start menu
shows the current game version) but not too often.